requirements.txt
# Core Dependencies
playwright==1.43.0
python-dotenv==1.0.1

# LangChain and LangGraph
langgraph==0.2.20
langchain==0.2.14
langchain-openai==0.1.22
langchain-google-genai==1.0.10
langchain-community==0.2.12

# AI/ML Libraries
openai==1.40.0
google-generativeai==0.7.2
pillow==10.4.0

# Web Scraping
scrapy==2.11.2
beautifulsoup4==4.12.3
lxml==5.3.0

# Data Processing
pandas==2.2.2
numpy==1.26.4

# Visualization and Dashboard
streamlit==1.37.1
plotly==5.23.0
matplotlib==3.9.2

# Utilities
aiohttp==3.10.5
requests==2.32.3
pydantic==2.8.2
tenacity==9.0.0

# Testing (optional)
pytest==8.3.2
pytest-asyncio==0.23.8
